requirements.txt
clickhouse-connect==0.6.20
requests==2.28.2
python-dotenv==1.0.0
backoff==2.2.1
ratelimit==2.2.1
tenacity==8.2.2
loguru==0.7.2